+++ src/argilla/tasks/users/create_default.py
@@ -37,7 +37,7 @@ def create_default(
             User(
                 first_name="",
                 username=DEFAULT_USERNAME,
-                role=UserRole.admin,
+                role=UserRole.owner,
                 api_key=api_key,
                 password_hash=accounts.hash_password(password),
                 workspaces=[Workspace(name=DEFAULT_USERNAME)],
@@ -49,6 +49,7 @@ def create_default(
             typer.echo("User with default credentials successfully created:")
             typer.echo(f"• username: {DEFAULT_USERNAME!r}")
             typer.echo(f"• password: {password!r}")
+            typer.echo(f"• role:     {UserRole.owner.value!r}")
             typer.echo(f"• api_key:  {api_key!r}")