+++ tests/server/api/v1/test_datasets.py
@@ -35,9 +35,6 @@ from argilla.server.schemas.v1.datasets import (
     DATASET_CREATE_GUIDELINES_MAX_LENGTH,
     FIELD_CREATE_NAME_MAX_LENGTH,
     FIELD_CREATE_TITLE_MAX_LENGTH,
-    LABEL_SELECTION_DESCRIPTION_MAX_LENGTH,
-    LABEL_SELECTION_TEXT_MAX_LENGTH,
-    LABEL_SELECTION_VALUE_MAX_LENGHT,
     QUESTION_CREATE_DESCRIPTION_MAX_LENGTH,
     QUESTION_CREATE_NAME_MAX_LENGTH,
     QUESTION_CREATE_TITLE_MAX_LENGTH,
@@ -45,6 +42,9 @@ from argilla.server.schemas.v1.datasets import (
     RATING_OPTIONS_MIN_ITEMS,
     RECORDS_CREATE_MAX_ITEMS,
     RECORDS_CREATE_MIN_ITEMS,
+    VALUE_TEXT_OPTION_DESCRIPTION_MAX_LENGTH,
+    VALUE_TEXT_OPTION_TEXT_MAX_LENGTH,
+    VALUE_TEXT_OPTION_VALUE_MAX_LENGTH,
     RecordInclude,
 )
 from argilla.server.search_engine import (
@@ -73,7 +73,6 @@ from tests.factories import (
     TextQuestionFactory,
     UserFactory,
     WorkspaceFactory,
-    WorkspaceUserFactory,
 )
 
 if TYPE_CHECKING:
@@ -1627,18 +1626,9 @@ async def test_create_dataset_field_with_nonexistent_dataset_id(
             {
                 "type": "label_selection",
                 "options": [
-                    {
-                        "value": "positive",
-                        "text": "Positive",
-                    },
-                    {
-                        "value": "negative",
-                        "text": "Negative",
-                    },
-                    {
-                        "value": "neutral",
-                        "text": "Neutral",
-                    },
+                    {"value": "positive", "text": "Positive"},
+                    {"value": "negative", "text": "Negative"},
+                    {"value": "neutral", "text": "Neutral"},
                 ],
             },
             {
@@ -1649,6 +1639,46 @@ async def test_create_dataset_field_with_nonexistent_dataset_id(
                     {"value": "neutral", "text": "Neutral", "description": None},
                 ],
                 "visible_options": None,
+            },
+        ),
+        (
+            {
+                "type": "ranking",
+                "options": [
+                    {"value": "completion-a", "text": "Completion A", "description": "Completion A is the best"},
+                    {"value": "completion-b", "text": "Completion B", "description": "Completion B is the best"},
+                    {"value": "completion-c", "text": "Completion C", "description": "Completion C is the best"},
+                    {"value": "completion-d", "text": "Completion D", "description": "Completion D is the best"},
+                ],
+            },
+            {
+                "type": "ranking",
+                "options": [
+                    {"value": "completion-a", "text": "Completion A", "description": "Completion A is the best"},
+                    {"value": "completion-b", "text": "Completion B", "description": "Completion B is the best"},
+                    {"value": "completion-c", "text": "Completion C", "description": "Completion C is the best"},
+                    {"value": "completion-d", "text": "Completion D", "description": "Completion D is the best"},
+                ],
+            },
+        ),
+        (
+            {
+                "type": "ranking",
+                "options": [
+                    {"value": "completion-a", "text": "Completion A", "description": None},
+                    {"value": "completion-b", "text": "Completion b", "description": None},
+                    {"value": "completion-c", "text": "Completion C", "description": None},
+                    {"value": "completion-d", "text": "Completion D", "description": None},
+                ],
+            },
+            {
+                "type": "ranking",
+                "options": [
+                    {"value": "completion-a", "text": "Completion A", "description": None},
+                    {"value": "completion-b", "text": "Completion b", "description": None},
+                    {"value": "completion-c", "text": "Completion C", "description": None},
+                    {"value": "completion-d", "text": "Completion D", "description": None},
+                ],
             },
         ),
     ],
@@ -1951,7 +1981,7 @@ async def test_create_dataset_question_with_nonexistent_dataset_id(
         {
             "type": "label_selection",
             "options": [
-                {"value": "".join(["a" for _ in range(LABEL_SELECTION_VALUE_MAX_LENGHT + 1)]), "text": "a"},
+                {"value": "".join(["a" for _ in range(VALUE_TEXT_OPTION_VALUE_MAX_LENGTH + 1)]), "text": "a"},
                 {"value": "b", "text": "b"},
             ],
         },
@@ -1962,7 +1992,7 @@ async def test_create_dataset_question_with_nonexistent_dataset_id(
         {
             "type": "label_selection",
             "options": [
-                {"value": "a", "text": "".join(["a" for _ in range(LABEL_SELECTION_TEXT_MAX_LENGTH + 1)])},
+                {"value": "a", "text": "".join(["a" for _ in range(VALUE_TEXT_OPTION_TEXT_MAX_LENGTH + 1)])},
                 {"value": "b", "text": "b"},
             ],
         },
@@ -1976,7 +2006,7 @@ async def test_create_dataset_question_with_nonexistent_dataset_id(
                 {
                     "value": "a",
                     "text": "a",
-                    "description": "".join(["a" for _ in range(LABEL_SELECTION_DESCRIPTION_MAX_LENGTH + 1)]),
+                    "description": "".join(["a" for _ in range(VALUE_TEXT_OPTION_DESCRIPTION_MAX_LENGTH + 1)]),
                 },
                 {"value": "b", "text": "b"},
             ],
@@ -1986,6 +2016,31 @@ async def test_create_dataset_question_with_nonexistent_dataset_id(
             "options": [
                 {"value": "a", "text": "a", "description": "a"},
                 {"value": "b", "text": "b", "description": "b"},
+                {"value": "b", "text": "b", "description": "b"},
+            ],
+        },
+        {
+            "type": "ranking",
+            "options": [
+                {"value": "a", "text": "a", "description": "a"},
+            ],
+        },
+        {
+            "type": "ranking",
+            "options": [
+                {"value": "a", "text": "a", "description": "a"},
+                {"value": "b", "text": "b", "description": "b"},
+                {"value": "b", "text": "b", "description": "b"},
+            ],
+        },
+        {
+            "type": "ranking",
+            "options": [
+                {
+                    "value": "a",
+                    "text": "a",
+                    "description": "".join(["a" for _ in range(VALUE_TEXT_OPTION_DESCRIPTION_MAX_LENGTH + 1)]),
+                },
                 {"value": "b", "text": "b", "description": "b"},
             ],
         },