+++ tests/server/api/v1/test_records.py
@@ -25,10 +25,10 @@ from sqlalchemy import func, select
 from sqlalchemy.orm import Session
 
 from tests.factories import (
-    AnnotatorFactory,
     DatasetFactory,
     LabelSelectionQuestionFactory,
     MultiLabelSelectionQuestionFactory,
+    RankingQuestionFactory,
     RatingQuestionFactory,
     RecordFactory,
     ResponseFactory,
@@ -62,6 +62,11 @@ async def create_multi_label_selection_questions(dataset: "Dataset") -> None:
         name="multi_label_selection_question_1", dataset=dataset, required=True
     )
     await MultiLabelSelectionQuestionFactory.create(name="multi_label_selection_question_2", dataset=dataset)
+
+
+def create_ranking_question(dataset: "Dataset") -> None:
+    RankingQuestionFactory.create(name="ranking_question_1", dataset=dataset, required=True)
+    RankingQuestionFactory.create(name="ranking_question_2", dataset=dataset)
 
 
 @pytest.mark.parametrize("response_status", ["submitted", "discarded", "draft"])
@@ -115,6 +120,20 @@ async def create_multi_label_selection_questions(dataset: "Dataset") -> None:
                 "values": {
                     "input_ok": {"value": "yes"},
                 },
+            },
+        ),
+        (
+            create_ranking_question,
+            {
+                "values": {
+                    "ranking_question_1": {
+                        "value": [
+                            {"value": "completion-b", "rank": 1},
+                            {"value": "completion-c", "rank": 2},
+                            {"value": "completion-a", "rank": 3},
+                        ]
+                    },
+                }
             },
         ),
     ],
@@ -216,6 +235,62 @@ async def test_create_submitted_record_response_with_missing_required_questions(
                 },
             },
         ),
+        (
+            create_ranking_question,
+            {
+                "values": {
+                    "ranking_question_2": {
+                        "value": [
+                            {"value": "completion-b", "rank": 1},
+                            {"value": "completion-c", "rank": 2},
+                            {"value": "completion-a", "rank": 3},
+                        ]
+                    },
+                }
+            },
+        ),
+        (
+            create_ranking_question,
+            {
+                "values": {
+                    "ranking_question_2": {
+                        "value": [
+                            {"value": "completion-b", "rank": 1},
+                            {"value": "completion-c", "rank": 1},
+                            {"value": "completion-a", "rank": 3},
+                        ]
+                    },
+                }
+            },
+        ),
+        (
+            create_ranking_question,
+            {
+                "values": {
+                    "ranking_question_2": {
+                        "value": [
+                            {"value": "completion-b", "rank": 1},
+                            {"value": "completion-c", "rank": 3},
+                            {"value": "completion-a", "rank": 3},
+                        ]
+                    },
+                }
+            },
+        ),
+        (
+            create_ranking_question,
+            {
+                "values": {
+                    "ranking_question_2": {
+                        "value": [
+                            {"value": "completion-b", "rank": 1},
+                            {"value": "completion-c", "rank": 1},
+                            {"value": "completion-a", "rank": 1},
+                        ]
+                    },
+                }
+            },
+        ),
     ],
 )
 @pytest.mark.asyncio
@@ -310,7 +385,7 @@ async def test_create_record_response_with_extra_question_responses(client: Test
                     "multi_label_selection_question_1": {"value": "wrong-type"},
                 },
             },
-            "Expected list of values, found <class 'str'>",
+            "This MultiLabelSelection question expects a list of values, found <class 'str'>",
         ),
         (
             create_multi_label_selection_questions,
@@ -319,12 +394,111 @@ async def test_create_record_response_with_extra_question_responses(client: Test
                     "multi_label_selection_question_1": {"value": ["option4", "option5"]},
                 },
             },
-            "['option4', 'option5'] are not valid options.\nValid options are: ['option1', 'option2', 'option3']",
+            "['option4', 'option5'] are not valid options for this MultiLabelSelection question.\nValid options are: ['option1', 'option2', 'option3']",
         ),
         (
             create_multi_label_selection_questions,
             {"values": {"multi_label_selection_question_1": {"value": []}}},
-            "Expected list of values, found empty list",
+            "This MultiLabelSelection question expects a list of values, found empty list",
+        ),
+        (
+            create_ranking_question,
+            {"values": {"ranking_question_1": {"value": "wrong-type"}}},
+            "This Ranking question expects a list of values, found <class 'str'>",
+        ),
+        (
+            create_ranking_question,
+            {"values": {"ranking_question_1": {"value": []}}},
+            "This Ranking question expects a list containing 3 values, found a list of 0 values",
+        ),
+        (
+            create_ranking_question,
+            {
+                "values": {
+                    "ranking_question_1": {
+                        "value": [
+                            {"value": "completion-b", "rank": 1},
+                        ]
+                    }
+                }
+            },
+            "This Ranking question expects a list containing 3 values, found a list of 1 values",
+        ),
+        (
+            create_ranking_question,
+            {
+                "values": {
+                    "ranking_question_1": {
+                        "value": [
+                            {"value": "completion-b", "rank": 1},
+                            {"value": "completion-c", "rank": 2},
+                            {"value": "completion-a", "rank": 3},
+                            {"value": "completion-z", "rank": 4},
+                        ],
+                    }
+                }
+            },
+            "This Ranking question expects a list containing 3 values, found a list of 4 values",
+        ),
+        (
+            create_ranking_question,
+            {
+                "values": {
+                    "ranking_question_1": {
+                        "value": [
+                            {"value": "completion-b", "rank": 1},
+                            {"value": "completion-c", "rank": 2},
+                            {"value": "completion-a", "rank": 4},
+                        ]
+                    }
+                }
+            },
+            "[4] are not valid ranks for this Ranking question.\nValid ranks are: [1, 2, 3]",
+        ),
+        (
+            create_ranking_question,
+            {
+                "values": {
+                    "ranking_question_1": {
+                        "value": [
+                            {"value": "completion-b"},
+                            {"value": "completion-c"},
+                            {"value": "completion-a"},
+                        ]
+                    }
+                }
+            },
+            "[None] are not valid ranks for this Ranking question.\nValid ranks are: [1, 2, 3]",
+        ),
+        (
+            create_ranking_question,
+            {
+                "values": {
+                    "ranking_question_1": {
+                        "value": [
+                            {"value": "completion-z", "rank": 1},
+                            {"value": "completion-c", "rank": 2},
+                            {"value": "completion-a", "rank": 3},
+                        ]
+                    }
+                }
+            },
+            "['completion-z'] are not valid options for this Ranking question.\nValid options are: ['completion-a', 'completion-b', 'completion-c']",
+        ),
+        (
+            create_ranking_question,
+            {
+                "values": {
+                    "ranking_question_1": {
+                        "value": [
+                            {"value": "completion-a", "rank": 1},
+                            {"value": "completion-c", "rank": 2},
+                            {"value": "completion-a", "rank": 3},
+                        ]
+                    }
+                }
+            },
+            "This Ranking question expects a list of unique values, but duplicates were found",
         ),
     ],
 )