+++ tests/tasks/users/test_create_default.py
@@ -32,7 +32,7 @@ def test_create_default(sync_db: "Session", cli_runner: CliRunner, cli: Typer):
 
     default_user = sync_db.query(User).filter_by(username=DEFAULT_USERNAME).first()
     assert default_user
-    assert default_user.role == UserRole.admin
+    assert default_user.role == UserRole.owner
     assert default_user.api_key == DEFAULT_API_KEY
     assert accounts.verify_password(DEFAULT_PASSWORD, default_user.password_hash)
     assert [ws.name for ws in default_user.workspaces] == [DEFAULT_USERNAME]
@@ -47,7 +47,7 @@ def test_create_default_with_specific_api_key_and_password(sync_db: "Session", c
 
     default_user = sync_db.query(User).filter_by(username=DEFAULT_USERNAME).first()
     assert default_user
-    assert default_user.role == UserRole.admin
+    assert default_user.role == UserRole.owner
     assert default_user.api_key == "my-api-key"
     assert accounts.verify_password("my-password", default_user.password_hash)
     assert [ws.name for ws in default_user.workspaces] == [DEFAULT_USERNAME]